requirements.txt
Package            Version
------------------ -----------
beautifulsoup4     4.12.3
certifi            2024.12.14
charset-normalizer 3.4.1
frozendict         2.4.6
html5lib           1.1
idna               3.10
lxml               5.3.0
multitasking       0.0.11
numpy              2.2.2
pandas             2.2.3
peewee             3.17.8
pip                24.3.1
platformdirs       4.3.6
python-dateutil    2.9.0.post0
python-dotenv      1.0.1
pytz               2024.2
requests           2.32.3
six                1.17.0
soupsieve          2.6
tzdata             2025.1
urllib3            2.3.0
webencodings       0.5.1
yfinance           0.2.52
